remove `GoToDefLocalVariableDeclaration`
